Scores C++
This lab was used as an introduction on opening a file and appending information from a text file into a file using an example array of average scores from students. This lab was meant to challenge us to take what we already know about 1D arrays and integrate it into our newfound knowledge of 2D arrays and opening/closing files. It made us think critically and analytically as we know the components of what to do and how to use it, but we had to figure out how to put it together and translate it onto a new concept introduced to us. It’s like taking puzzle pieces, putting them together then using the whole picture to connect the dots with the other puzzles you’ve learned to put together, that’s how my partner and I viewed it as. We knew how to use 1D arrays, we knew how to get the average, the highest average, and the lowest average which are the puzzles we had already put together but the challenge was compiling it into one code that translates it into a file which is the part where we connect the dots. Which is the overall purpose of the lab in terms of introducing files into our previous knowledge of arrays and using that to compile one whole program that challenges us intellectually and in terms of problem solving.